third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-step/debugger-step-out-custom-element-callbacks.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function createdCallback()
    {
        debugger;
        output('Invoked createdCallback.');
    };
    proto.attachedCallback = function attachedCallback()
    {
        output('Invoked attachedCallback.');
    };
    proto.detachedCallback = function detachedCallback()
    {
        output('Invoked detachedCallback.');
    };
    proto.attributeChangedCallback = function attributeChangedCallback()
    {
        output('Invoked attributeChangedCallback.');
    };
    var FooElement = document.registerElement('x-foo', { prototype: proto });
    var foo = new FooElement();
    foo.setAttribute('a', 'b');
    document.body.appendChild(foo);
    foo.remove();
}

function test() {
  SourcesTestRunner.startDebuggerTest(step1, true);

  function step1() {
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(step2);
  }

  function step2() {
    var actions = [
      'Print',              // debugger; in createdCallback
      'StepOut', 'Print',   // at foo.setAttribute()
      'StepInto', 'Print',  // at attributeChangedCallback
      'StepOut', 'Print',   // at document.body.appendChild()
      'StepInto', 'Print',  // at attachedCallback
      'StepOut', 'Print',   // at foo.remove()
      'StepInto', 'Print',  // at detachedCallback
      'StepOut', 'Print',   // at testFunction() return point
    ];
    SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions, step3);
  }

  function step3() {
    SourcesTestRunner.completeDebuggerTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests stepping out from custom element callbacks.
</p>
</body>
</html>
