third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-step/debugger-step-out-custom-element-callbacks-expected.txt

Tests stepping out from custom element callbacks.

[page] Invoked createdCallback.
[page] Invoked attributeChangedCallback.
[page] Invoked attachedCallback.
[page] Invoked detachedCallback.
Set timer for test function.
Call stack:
    0) createdCallback (debugger-step-out-custom-element-callbacks.html:12)
    1) testFunction (debugger-step-out-custom-element-callbacks.html:28)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOut...
Call stack:
    0) testFunction (debugger-step-out-custom-element-callbacks.html:29)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepInto...
Call stack:
    0) attributeChangedCallback (debugger-step-out-custom-element-callbacks.html:25)
    1) testFunction (debugger-step-out-custom-element-callbacks.html:29)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOut...
Call stack:
    0) testFunction (debugger-step-out-custom-element-callbacks.html:30)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepInto...
Call stack:
    0) attachedCallback (debugger-step-out-custom-element-callbacks.html:17)
    1) testFunction (debugger-step-out-custom-element-callbacks.html:30)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOut...
Call stack:
    0) testFunction (debugger-step-out-custom-element-callbacks.html:31)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepInto...
Call stack:
    0) detachedCallback (debugger-step-out-custom-element-callbacks.html:21)
    1) testFunction (debugger-step-out-custom-element-callbacks.html:31)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Executing StepOut...
Call stack:
    0) testFunction (debugger-step-out-custom-element-callbacks.html:32)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
